Distinguish _subscript_ from _index_.
A _subscript_ is an unsigned integer, describing the distance from the beginning of an array to an element in it.
An _index_ is a signed integer, describing the distance from the beginning of an array to an element in it if it is non-negative, or the distance from the past-the-end position if it is negative.